[Fixbug] Use int64 as the output of arg-reduce (#111)
* use int64 as the output of arg-reduce

* .

+++ gallery/how-to-guides/add-new-operator-compute-definition.py
@@ -371,7 +371,7 @@ def arg_max_example():
         ),
     )
     task = Task('arg_max', inputs=[a], outputs=[b])
-    run_task(task, [hidet.randn([4, 3])], [hidet.empty([4], dtype='int32')])
+    run_task(task, [hidet.randn([4, 3])], [hidet.empty([4], dtype='int64')])
 
 
 arg_max_example()
